data/wow-workflows/pyproject.toml
[tool.poetry]
name = "wow-workflows"
version = "0.1.0"
description = ""
authors = ["Victor Tantillo <v.tantillo3@gmail.com>"]
readme = "README.md"
packages = [{include = "wow_workflows"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
prefect = "^2.8.7"
requests-toolbelt = "^0.10.1"
requests-oauthlib = "^1.3.1"
requests = "^2.28.2"
pandas = "^2.0.0"
pyjanitor = "^0.24.0"
polars = "^0.16.18"
pydantic = "^1.10.7"
datamodel-code-generator = "^0.17.2"
ulid-py = "^1.1.0"
databases = {extras = ["asycpg"], version = "^0.7.0"}
psycopg = {extras = ["binary"], version = "^3.1.8"}


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
black = "^23.3.0"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
